fix: adress pr comments
